More playground polish
