+++ website/src/playground/PlaygroundLoader.tsx
@@ -1,6 +1,28 @@
-import { useEffect, useState, useRef } from "react";
-import { LoadingState } from "./types";
-import { getCurrentCode, getFileState, usePlaygroundState } from "./utils";
+import { useEffect, useState, useRef, SetStateAction, Dispatch } from "react";
+import {
+	defaultPlaygroundState,
+	emptyPrettierOutput,
+	emptyRomeOutput,
+	IndentStyle,
+	LoadingState,
+	PlaygroundSettings,
+	PlaygroundState,
+	QuoteProperties,
+	QuoteStyle,
+	TrailingComma,
+} from "./types";
+import {
+	createLocalStorage,
+	decodeCode,
+	encodeCode,
+	getCurrentCode,
+	getExtension,
+	getFileState,
+	isJSXFilename,
+	isScriptFilename,
+	isTypeScriptFilename,
+	normalizeFilename,
+} from "./utils";
 import Playground from "./Playground";
 import LoadingScreen from "./components/LoadingScreen";
 
@@ -151,6 +173,171 @@ function App() {
 				/>
 			);
 	}
+}
+
+const lastSearchStore = createLocalStorage("last-search");
+
+const FILE_QUERY_KEY_REGEX = /^files\.(.*?)$/;
+
+export function usePlaygroundState(): [
+	PlaygroundState,
+	Dispatch<SetStateAction<PlaygroundState>>,
+	() => void,
+] {
+	const searchQuery =
+		window.location.search === ""
+			? lastSearchStore.get() ?? ""
+			: window.location.search;
+
+	function initState(
+		searchParams: URLSearchParams,
+		ignoreFiles: boolean,
+	): PlaygroundState {
+		let singleFileMode = true;
+		let hasFiles = false;
+		let files: PlaygroundState["files"] = {};
+
+		if (!ignoreFiles) {
+			// Populate files
+			for (const [key, value] of searchParams) {
+				const match = key.match(FILE_QUERY_KEY_REGEX);
+				if (match != null) {
+					const filename = normalizeFilename(match[1]!);
+					files[filename] = {
+						content: decodeCode(value),
+						rome: emptyRomeOutput,
+						prettier: emptyPrettierOutput,
+					};
+					singleFileMode = false;
+					hasFiles = true;
+				}
+			}
+
+			// Single file mode
+			if (searchParams.get("code")) {
+				const ext = getExtension({
+					typescript: searchParams.get("typescript") === "true",
+					jsx: searchParams.get("jsx") === "true",
+					script: searchParams.get("script") === "true",
+				});
+				files[`main.${ext}`] = {
+					content: decodeCode(searchParams.get("code") ?? ""),
+					rome: emptyRomeOutput,
+					prettier: emptyPrettierOutput,
+				};
+				hasFiles = true;
+			}
+		}
+
+		if (!hasFiles) {
+			files = defaultPlaygroundState.files;
+		}
+
+		return {
+			cursorPosition: 0,
+			tab: searchParams.get("tab") ?? "formatter",
+			singleFileMode,
+			currentFile: Object.keys(files)[0] ?? "main.js",
+			files,
+			settings: {
+				lineWidth: parseInt(
+					searchParams.get("lineWidth") ??
+						String(defaultPlaygroundState.settings.lineWidth),
+				),
+				indentStyle:
+					(searchParams.get("indentStyle") as IndentStyle) ??
+					defaultPlaygroundState.settings.indentStyle,
+				quoteStyle:
+					(searchParams.get("quoteStyle") as QuoteStyle) ??
+					defaultPlaygroundState.settings.quoteStyle,
+				quoteProperties:
+					(searchParams.get("quoteProperties") as QuoteProperties) ??
+					defaultPlaygroundState.settings.quoteProperties,
+				trailingComma:
+					(searchParams.get("trailingComma") as TrailingComma) ??
+					defaultPlaygroundState.settings.trailingComma,
+				indentWidth: parseInt(
+					searchParams.get("indentWidth") ??
+						String(defaultPlaygroundState.settings.indentWidth),
+				),
+				enabledNurseryRules:
+					searchParams.get("enabledNurseryRules") === "true" ||
+					defaultPlaygroundState.settings.enabledNurseryRules,
+				enabledLinting:
+					searchParams.get("enabledLinting") === "true" ||
+					defaultPlaygroundState.settings.enabledLinting,
+			},
+		};
+	}
+
+	const [playgroundState, setPlaygroundState] = useState(() =>
+		initState(new URLSearchParams(searchQuery), window.location.search === ""),
+	);
+
+	function resetPlaygroundState() {
+		setPlaygroundState(initState(new URLSearchParams(""), false));
+	}
+
+	useEffect(() => {
+		const rawQueryParams: Record<string, unknown> = {
+			...playgroundState.settings,
+		};
+
+		// Eliminate default values
+		const queryStringObj: Record<string, string> = {};
+		for (const key in rawQueryParams) {
+			const defaultValue = String(
+				defaultPlaygroundState.settings[key as keyof PlaygroundSettings],
+			);
+			const rawValue = rawQueryParams[key];
+			const value = String(rawValue);
+
+			if (rawValue !== undefined && value !== defaultValue) {
+				queryStringObj[key] = value;
+			}
+		}
+
+		if (
+			playgroundState.singleFileMode &&
+			Object.keys(playgroundState.files).length === 1
+		) {
+			// Single file mode
+			const code = getCurrentCode(playgroundState);
+			if (code) {
+				queryStringObj.code = encodeCode(code);
+			}
+
+			if (!isTypeScriptFilename(playgroundState.currentFile)) {
+				queryStringObj.typescript = "false";
+			}
+
+			if (!isJSXFilename(playgroundState.currentFile)) {
+				queryStringObj.jsx = "false";
+			}
+
+			if (isScriptFilename(playgroundState.currentFile)) {
+				queryStringObj.script = "true";
+			}
+		} else {
+			// Populate files
+			for (const filename in playgroundState.files) {
+				const content = playgroundState.files[filename]?.content ?? "";
+				queryStringObj[`files.${filename}`] = encodeCode(content);
+			}
+		}
+
+		const queryString = new URLSearchParams(queryStringObj).toString();
+		lastSearchStore.set(queryString);
+
+		let url = `${window.location.protocol}//${window.location.host}${window.location.pathname}`;
+		if (queryString !== "") {
+			url += `?${queryString}`;
+		}
+
+		window.history.replaceState({ path: url }, "", url);
+	}, [playgroundState]);
+
+	return [playgroundState, setPlaygroundState, resetPlaygroundState];
 }
 
 export default App;
