+++ website/src/playground/styles/_results.scss
@@ -14,7 +14,6 @@
 }
 
 .mermaid {
-  border-top: 1px solid var(--playground-border);
   display: flex;
   width: 100%;
   height: 100%;
