+++ website/src/playground/styles/_settings.scss
@@ -116,6 +116,7 @@
     padding: 2px 6px;
     color: inherit;
     border-radius: 6px;
+    cursor: pointer;
 
     @include dark-mode {
       border-color: rgba(255, 255, 255, 0.1);
@@ -172,7 +173,6 @@
   width: 100%;
   height: 100%;
   background-color: rgba(255, 255, 255, 0.1) !important;
-  border-top: 1px solid var(--playground-border);
 }
 
 .file-view {
@@ -188,20 +188,44 @@
   }
 
   > input[type=text] {
-    width: 100%;
+    display: block;
     padding: 5px;
     margin: 5px;
+    margin-bottom: 0;
   }
 
   .files-list {
     padding-top: 10px;
 
     li {
-      padding: 4px 10px;
+      display: flex;
+      align-items: center;
+      padding: 0 10px;
+      cursor: pointer;
+      line-height: 30px;
+      height: 30px;
+
+      &:hover {
+        background: rgba(#000, 5%);
+
+        button {
+          display: block;
+          opacity: 0.8;
+        }
+      }
 
       &.active {
         font-weight: bold;
         background: rgba(#000, 10%);
+      }
+
+      button {
+        display: none;
+        margin-left: auto;
+
+        &:hover {
+          opacity: 1;
+        }
       }
     }
   }
