+++ website/src/playground/styles/index.scss
@@ -53,7 +53,11 @@ h2 {
   font-size: 14px;
   font-weight: normal;
   box-shadow: inset 0 1px 0 var(--hard-border-color), inset 0 -1px 0 var(--hard-border-color);
-  background-color: rgba(255, 255, 255, 0.15);
+  background-color: rgba(#000, 0.05);
+
+  @include dark-mode {
+    background-color: rgba(#fff, 0.15);
+  }
 
   &.rome {
     background-color: rgba($color-4, 30%);
